added toolbar to layouts, consisting of filter, sort, and rangeSelector components which were newly organized outside of marketplace. Props can be used to specify which components should be included in toolbar layout. Styles were extracted from marketplace.less, and placed within individual less files for each component
