+++ packages/client/src/components/range-selector/assets/rangeSelector.less
@@ -1,0 +1,20 @@
+.rangeSelectors {
+    display: flex;
+    height: 25px;
+    width: 120px;
+    justify-content: space-around;
+    align-items: center;
+    button.rangeSelector {
+        height: 26px;
+        cursor: pointer;
+        border: none;
+        background: none;
+        color: grey;
+        width: 30%;
+    }
+    button.rangeSelector.selected {
+        color: #6065CB;
+        font-weight: 600;
+        box-shadow: none;
+    }
+}
