+++ packages/client/src/layouts/toolbar/assets/toolbar.less
@@ -1,0 +1,7 @@
+.toolbar {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    height: 30px;
+    margin-top: 20px;
+}
